\name{BWR}
\alias{BWR}
 
\title{
BW ratio calculation
}
\description{
Compute the BW ratios from data groups
}
\usage{
BWR(X,
    groupLabel)
}
 
\arguments{
  \item{X}{
The spectral dataset in the matrix format in which each row contains a single sample
}
  \item{groupLabel}{
Group label of samples in the dataset.
}
}
\details{
Compute the BW ratios from data groups 
}
\value{
Return BW ratio 
}
\author{
Trung Nghia VU 
}

\seealso{
\code{\link{createNullSampling}}    
}
\examples{ cat("\n Please see more examples in the vignettes file.")

}
 
 
\keyword{ BW }

