% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWinedata.subset.R
\name{GetWinedata.subset}
\alias{GetWinedata.subset}
\title{Get subset of Winedata for code examples}
\usage{
GetWinedata.subset()
}
\value{
list of 2: spectra, ppm values, color and origin.
}
\description{
This functions extracts a small part of the Winedata to be used in code testing and code examples
}
\examples{

subset <- GetWinedata.subset()
subset.spectra = subset$Spectra
subset.ppm = subset$PPM.vector

}
\author{
Charlie Beirnaert, \email{charlie.beirnaert@uantwerpen.be}
}
