\name{speaq-package}
\alias{speaq-package}
\alias{speaq}
\docType{package}
\title{
An R-package for NMR spectrum alignment and quantitation
}
\description{
Integrated NMR spectrum alignment and quantitation based on Hierarchial Clustering Based Peak Alignment (CluPA)}
\details{
\tabular{ll}{
Package: \tab speaq\cr
Type: \tab Package\cr
Version: \tab 1.2.0\cr
Date: \tab 2015-05-01\cr
License: \tab Apache License 2.0\cr
LazyLoad: \tab yes\cr
}


}
\author{
Trung Nghia VU, Kris Laukens and Dirk Valkenborg

Maintainer: Trung Nghia VU <nghiavtr@gmail.com>

}
\references{
Vu TN, Valkenborg D, Smets K, Verwaest KA, Dommisse R, Lemie`re F, Verschoren A, Goethals B, Laukens K. (2011) An integrated workflow for robust alignment and simplified quantitative analysis of NMR spectrometry data. BMC Bioinformatics. 2011 Oct 20;12:405.
}

\keyword{ package }


