% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Winedata.R
\docType{data}
\name{Winedata}
\alias{Winedata}
\title{Wine dataset}
\format{
A list with the spectra, ppm values, color and origin as list entries.
}
\source{
University of Copenhagen, Dept. of Food Science, Quality & Technology. Available at 'models.life.ku.dk/datasets'
}
\usage{
data(Winedata)
}
\description{
1H-NMR data of 40 wines, different origins and colors are included.
}
\examples{
data(Winedata)
Spectra <- Winedata$spectra 
ppm.wine <- Winedata$ppm 
wine.color <- Winedata$wine.color 
wine.origin <- Winedata$origin 
}
\references{
Larsen et al. (2006) An exploratory chemometric study of 1H-NMR spectra of table wines. J.Chemom. 20 (2006) 198-208
(\href{https://onlinelibrary.wiley.com/doi/10.1002/cem.991/full}{Wiley Online Library})
}
\keyword{datasets}
